pyproject.toml
[project]
name = "book-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "rukhsanagull", email = "160248041+misbahkhan99@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.108.0",
    "crewai-tools>=0.38.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
]

[project.scripts]
book-crew = "book_crew:main"
book = "book_crew.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
